requirements.txt
Flask==2.0.1
python-telegram-bot==13.7
gunicorn==20.1.0
requests==2.26.0
beautifulsoup4==4.9.3
python-dotenv==0.19.0
Flask-SQLAlchemy==2.5.1
tmdbv3api==1.7.7
